% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mittelwert}
\alias{mittelwert}
\title{Mittelwertfunktion}
\usage{
mittelwert(...)
}
\arguments{
\item{...}{same arguments as the mean function}
}
\value{
arithmetisches Mittel
}
\description{
Mittelwertfunktion
}
