% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{oeffne_in_excel}
\alias{oeffne_in_excel}
\title{Datensatz in Excel öffnen}
\usage{
oeffne_in_excel(daten)
}
\arguments{
\item{daten}{Datensatz mit den Daten, die in Excel gezeigt werden sollen}
}
\description{
Datensatz in Excel öffnen
}
