% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{standardabweichung}
\alias{standardabweichung}
\title{Standardabweichung (unkorrigiert)}
\usage{
standardabweichung(daten)
}
\arguments{
\item{daten}{Daten zur Berechnung}
}
\value{
unkorrigierte Standardabweichung
}
\description{
Standardabweichung (unkorrigiert)
}
